Fix Document.version (#168)
## What does this PR do ?

We always assume the `version` attribute of a document would be a long but users can set it to, for exemple, a string.
Because the SDK 4 is coming and to avoid a breaking change, we now ignore the version attribute if it is not a long to set in the Document class as if no version was given. The version still remain in the content of the document.